Create last_update_export.py
This script is designed for use in a "disconnect" environment. The host server is to do daily exports based on last build of a channel. This script looks at the list of channels, gets the ChannelLastBuildByID and checks if the date is the current date or prior and then exports that channel.

Logging is provided as well to show which channels were updated in that export. 

The target server can then rsync the `/mnt/export` directory and run another script to execute the inter-server-sync import of the exported data.